feat: added support for custom object inspectors - added Utils.addInspector method for registering custom inspectors - moved existing `inspect()` function to `defaultInspector.js`, and replaced the `inspect()` function with logic for determining the string representation from any available custom inspectors - added unit tests for above
