+++ lib/chai/utils/index.js
@@ -100,6 +100,12 @@ exports.addProperty = require('./addProperty');
 exports.addMethod = require('./addMethod');
 
 /*!
+ * add Inspector
+ */
+
+exports.addInspector = require('./addInspector');
+
+/*!
  * overwrite Property
  */
