+++ lib/chai/utils/inspectors.js
@@ -1,0 +1,9 @@
+/**
+ * @typedef {function(obj: *): string} Inspector
+ */
+
+/**
+ * Stores an array of custom inspectors
+ * @type {Inspector[]}
+ */
+module.exports = [];
